+++ app/src/main/java/me/saket/dank/ui/media/MediaAlbumViewerActivity.java
@@ -248,7 +248,7 @@ public class MediaAlbumViewerActivity extends DankActivity implements MediaFragm
             optimizedQualityUrl = activeMediaItem.mediaLink().lowQualityUrl();
           } else {
             ImageWithMultipleVariants imageVariants = ImageWithMultipleVariants.Companion.of(redditSuppliedImages);
-            optimizedQualityUrl = imageVariants.findNearestFor(
+            optimizedQualityUrl = imageVariants.findNearestUrlFor(
                 getResources().getDisplayMetrics().widthPixels,
                 ImageWithMultipleVariants.DEFAULT_VIEWER_MIN_WIDTH,
                 activeMediaItem.mediaLink().lowQualityUrl() /* defaultValue */
@@ -649,7 +649,7 @@ public class MediaAlbumViewerActivity extends DankActivity implements MediaFragm
     Observable<File> optimizedResImageFileStream = getRedditSuppliedImages()
         .flatMapObservable(redditImages -> Observable.create(emitter -> {
           ImageWithMultipleVariants imageVariants = ImageWithMultipleVariants.Companion.of(redditImages);
-          String optimizedQualityImageForDevice = imageVariants.findNearestFor(
+          String optimizedQualityImageForDevice = imageVariants.findNearestUrlFor(
               getDeviceDisplayWidth(),
               ImageWithMultipleVariants.DEFAULT_VIEWER_MIN_WIDTH,
               albumItem.mediaLink().lowQualityUrl()