+++ app/src/main/java/me/saket/dank/ui/subreddit/uimodels/SubredditSubmission.java
@@ -15,6 +15,7 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions;
 import com.bumptech.glide.request.RequestOptions;
+import com.bumptech.glide.signature.ObjectKey;
 import com.google.auto.value.AutoValue;
 import com.jakewharton.rxrelay2.PublishRelay;
 
@@ -132,6 +133,8 @@ public interface SubredditSubmission {
 
       public abstract String contentDescription();
 
+      public abstract Optional<Integer> height();
+
       public static Builder builder() {
         return new AutoValue_SubredditSubmission_UiModel_Thumbnail.Builder();
       }
@@ -150,6 +153,8 @@ public interface SubredditSubmission {
 
         public abstract Builder contentDescription(String description);
 
+        public abstract Builder height(Optional<Integer> height);
+
         public abstract Thumbnail build();
       }
     }
@@ -164,6 +169,7 @@ public interface SubredditSubmission {
     private UiModel uiModel;
     private final ConstraintSet originalConstraintSet = new ConstraintSet();
     private final ConstraintSet leftAlignedThumbnailConstraintSet = new ConstraintSet();
+    private final int largeImageHeight;
 
     protected ViewHolder(View itemView) {
       super(itemView);
@@ -180,6 +186,8 @@ public interface SubredditSubmission {
       int thumbnailGoneMargin = itemView.getContext().getResources().getDimensionPixelSize(R.dimen.subreddit_submission_padding);
       leftAlignedThumbnailConstraintSet.setGoneMargin(R.id.submission_item_title, ConstraintSet.START, thumbnailGoneMargin);
       originalConstraintSet.setGoneMargin(R.id.submission_item_title, ConstraintSet.END, thumbnailGoneMargin);
+
+      largeImageHeight = itemView.getContext().getResources().getDimensionPixelSize(R.dimen.subreddit_submission_large_image);
     }
 
     public void setUiModel(UiModel uiModel) {
@@ -273,23 +281,34 @@ public interface SubredditSubmission {
               break;
             case THUMBNAIL:
               imageView.setVisibility(View.GONE);
-              Glide.with(itemView)
-                  .load(thumb.remoteUrl().get())
-                  .apply(RequestOptions.circleCropTransform())
-                  .transition(DrawableTransitionOptions.withCrossFade())
-                  .into(thumbnailView);
+              loadThumbnail(thumb, RequestOptions.circleCropTransform(), thumbnailView, -1);
               break;
             case LARGE:
-              thumbnailView.setVisibility(View.GONE);
-              Glide.with(itemView)
-                  .load(thumb.remoteUrl().get())
-                  .apply(RequestOptions.centerCropTransform())
-                  .transition(DrawableTransitionOptions.withCrossFade())
-                  .into(imageView);
+              setLargeImage(thumb, largeImageHeight);
+              break;
+            case FULL_HEIGHT:
+              setLargeImage(thumb, thumb.height().orElse(ViewGroup.LayoutParams.WRAP_CONTENT));
               break;
           }
         }
       });
+    }
+
+    private void setLargeImage(UiModel.Thumbnail thumb, int height) {
+      ViewGroup.LayoutParams layoutParams = imageView.getLayoutParams();
+      layoutParams.height = height;
+      imageView.setLayoutParams(layoutParams);
+      thumbnailView.setVisibility(View.GONE);
+      loadThumbnail(thumb, RequestOptions.centerCropTransform(), imageView, height);
+    }
+
+    private void loadThumbnail(UiModel.Thumbnail thumb, RequestOptions requestOptions, ImageView imageView, int height) {
+      Glide.with(itemView)
+          .load(thumb.remoteUrl().get())
+          .signature(new ObjectKey(height))
+          .apply(requestOptions)
+          .transition(DrawableTransitionOptions.withCrossFade())
+          .into(imageView);
     }
 
     @Override